requirements.txt
asgiref==3.8.0
Django==5.0.3
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.0
pip==24.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
setuptools==69.2.0
shortuuid==1.0.13
sqlparse==0.4.4
typing_extensions==4.10.0
wheel==0.43.0
whitenoise==6.6.0
